+++ src/components/Navigation/NavigationMenu/utils/utils.ts
@@ -1,0 +1,9 @@
+export const sanitizePathname = (pathname: string) => {
+  if (pathname[0] === "/") {
+    return String(pathname).split("/")[1];
+  }
+
+  return String(pathname);
+};
+
+export default sanitizePathname;