+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v1.42.3
-- **[fix]** - `DialogTitle` padding with `DialogCloseIcon` padding
-- 
+## v1.43.0
+- **[feat]** - Add option for hide mobile menu on scroll
+- **[feat]** - Add toggle mobile menu when is open
+- **[feat]** - Add option for diable mobile menu